[feature-55] add config for per page and move the func to presenter
