+++ conf/app.conf
@@ -4,6 +4,7 @@ httpport = ${PORT||8080}
 runmode = ${APP_RUN_MODE||dev}
 db_url = ${DATABASE_URL}
 SessionOn = true
+PaginationPerPage = 10
 
 [dev]
 SessionProvider = postgresql
