+++ helpers/helpers_suite_test.go
@@ -3,8 +3,6 @@ package helpers_test
 import (
 	"testing"
 
-	"go-google-scraper-challenge/initializers"
-
 	. "github.com/onsi/ginkgo"
 	. "github.com/onsi/gomega"
 )
@@ -13,7 +11,3 @@ func TestHelpers(t *testing.T) {
 	RegisterFailHandler(Fail)
 	RunSpecs(t, "Helpers Suite")
 }
-
-var _ = BeforeSuite(func() {
-	initializers.SetupTestEnvironment()
-})
